LayoutTests/fast/css-grid-layout/mark-as-infinitely-growable-expected.txt

PASS window.getComputedStyle(gridAutoAndAuto, '').getPropertyValue('grid-template-columns') is "30px 70px"
PASS window.getComputedStyle(gridAutoAndAuto, '').getPropertyValue('grid-template-rows') is "10px"
PASS window.getComputedStyle(gridFixedMaxContentAndFixedMaxContent, '').getPropertyValue('grid-template-columns') is "20px 80px"
PASS window.getComputedStyle(gridFixedMaxContentAndFixedMaxContent, '').getPropertyValue('grid-template-rows') is "10px"
PASS successfullyParsed is true

TEST COMPLETE
X X
XXXXXX XXX
XX
XXXX X XXX
